[sf-0003] change the name of the project. Rename all the important files of the project and decided a basic design
