+++ BE-SmartFenneck/local_config/upload-path.ts
@@ -1,0 +1,1 @@
+export const uploadPath = './upload';
