+++ FE-SmartFenneck/src/app/api-routes/enums/api-route.ts
