+++ FE-SmartFenneck/src/app/api-routes/utils/generate-api-route.ts
