+++ FE-SmartFenneck/src/app/app.component.html
@@ -1,0 +1,5 @@
+<router-outlet></router-outlet>
+<sf-header></sf-header>
+<section class="app-content">
+    <sf-upload (fileSelection)="onFileSelected($event)"></sf-upload>
+</section>
