+++ FE-SmartFenneck/src/app/app.component.scss
@@ -2,5 +2,4 @@
 
 :host {
     display: block;
-    background-color: $sf-white-sand;
 }
