+++ FE-SmartFenneck/src/app/app.component.spec.ts
@@ -23,13 +23,13 @@ describe('AppComponent', () => {
   it(`should have as title 'SmartFileFE'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('SmartFileFE');
+    expect(app.title).toEqual('FE-SmartFenneck');
   });
 
   it('should render title in a h1 tag', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to SmartFileFE!');
+    expect(compiled.querySelector('h1').textContent).toContain('Welcome to FE-SmartFenneck!');
   });
 });
