+++ FE-SmartFenneck/src/app/app.component.ts
