+++ FE-SmartFenneck/src/app/components/header/header.component.html
@@ -1,5 +1,5 @@
 <header>
     <div class="wrapper">
-        <img [src]="'../assets/smartfile.svg'"/>
+        <img [src]="'../assets/fennec.svg'"/>
     </div>
 </header>
