+++ FE-SmartFenneck/src/core-scss/color-variables.scss
@@ -4,3 +4,5 @@ $sf-light-grey: #eaeaea;
 $sf-tiny-sweet-blue: #b5e9e9;
 $sf-persian-blue: #00395a;
 $sf-white-sand: #f4f4f4;
+$sf-fennec-color: #ec4d37;
+$sf-burnt-sienna: #ef5455;
