+++ FE-SmartFenneck/src/styles.scss
@@ -3,5 +3,6 @@
 html, body {
     margin: 0;
     padding: 0;
-    background-color: $sf-white-sand;
+    background-color: ghostwhite;
+    font-family: 'Roboto', sans-serif;
 }
