--- SmartFileFE/src/app/app.component.html
@@ -1,3 +1,0 @@
-<sf-header></sf-header>
-<sf-upload (fileSelection)="onFileSelected($event)"></sf-upload>
-<router-outlet></router-outlet>
