--- SmartFileFE/src/app/components/header/header.component.scss
@@ -1,29 +1,0 @@
-@import '../../../core-scss/mixins';
-@import '../../../core-scss/color-variables';
-
-$height: vw(60);
-
-:host {
-    background-color: $sf-persian-blue;
-    height: $height;
-    display: block;
-    
-    header {
-        position: relative;
-        display: block;
-        width: 100%;
-        height: $height;
-        padding-left: vw(5);
-        
-        .wrapper {
-            position: absolute;
-            display: block;
-            width: vw(60);
-            height: $height;
-            
-            img {
-                width: 100%;
-            }
-        }
-    }
-}
