+++ app.ts
@@ -5,8 +5,8 @@ import * as bodyParser from 'body-parser';
 import * as cors from 'cors';
 import * as path from 'path';
 
-import {AppRoutes} from './SmartFileBE/routes/app-routes/app.routes';
-import {FileController} from './SmartFileBE/controllers/file';
+import {AppRoutes} from './BE-SmartFenneck/routes/app-routes/app.routes';
+import {FileController} from './BE-SmartFenneck/controllers/file';
 
 
 class App {
