+++ nodemon.json
@@ -1,6 +1,6 @@
 {
     "ignore": ["**/*.test.ts", "**/*.spec.ts", ".git", "node_modules"],
-    "watch": ["SmartFileBE", "app.ts", "index.ts"],
+    "watch": ["BE-SmartFenneck", "app.ts", "index.ts"],
     "exec": "npm start",
     "ext": "ts"
 }
